Solución errores selección de cargos

+++ config/cargos.php
@@ -1,5 +1,7 @@
 <?php
 require('../bootstrap.php');
+error_reporting(E_ALL);
+ini_set('display_errors', true);
 
 acl_acceso($_SESSION['cargo'], array('z', '1'));
 
@@ -12,7 +14,7 @@ include ("../menu.php");
 	</div>
 
 	<?php
-	if ($_GET['borrar']=='1') {
+	if (isset($_GET['borrar']) && $_GET['borrar'] == 1) {
 		$dni_profe = preg_replace('([^A-Za-z0-9])', '', $_GET['dni_profe']);
 		mysqli_query($db_con, "delete from departamentos where dni = '".$dni_profe."'");
 		echo '<div class="alert alert-success">
@@ -20,62 +22,62 @@ include ("../menu.php");
 	            El profesor ha sido borrado de la base de datos..
 	          </div>';
 	}
+
 	if (isset($_POST['enviar'])) {
 
-	// Backup de FTUTORES
-	mysqli_query($db_con, "drop table FTUTORES_seg" );
-	mysqli_query($db_con, "create table FTUTORES_seg select * from FTUTORES" );
-	mysqli_query($db_con, "truncate table FTUTORES" );
-	// Backup de departamentos
-	mysqli_query($db_con, "drop table departamentos_seg" );
-	mysqli_query($db_con, "create table departamentos_seg select * from departamentos" );
-	// Backup de c_profes
-	mysqli_query($db_con, "drop table c_profes_seg" );
-	mysqli_query($db_con, "create table c_profes_seg select * from c_profes" );
-	mysqli_query($db_con, "update c_profes set telefono = ''" );
+		// Backup de FTUTORES
+		mysqli_query($db_con, "drop table FTUTORES_seg" );
+		mysqli_query($db_con, "create table FTUTORES_seg select * from FTUTORES" );
+		mysqli_query($db_con, "truncate table FTUTORES" );
+		// Backup de departamentos
+		mysqli_query($db_con, "drop table departamentos_seg" );
+		mysqli_query($db_con, "create table departamentos_seg select * from departamentos" );
+		// Backup de c_profes
+		mysqli_query($db_con, "drop table c_profes_seg" );
+		mysqli_query($db_con, "create table c_profes_seg select * from c_profes" );
+		mysqli_query($db_con, "update c_profes set telefono = ''" );
 
-	mysqli_query($db_con, "truncate table cargos" );
+		mysqli_query($db_con, "truncate table cargos" );
 
 
-		foreach ( $_POST as $dni => $cargo_profe ) {
-			// echo "$dni => $cargo_profe<br>";
-			if ($cargo_profe == "Enviar") {
-				continue;
-			}
-			elseif (is_numeric($cargo_profe) and strlen ( $cargo_profe ) == "9") {
-				$dni = substr ( $dni, 1, -2 );
-				$n_profe = mysqli_query($db_con, "update c_profes set telefono = '$cargo_profe' where dni like '%$dni'" );
-				//echo "update c_profes set telefono = '$cargo_profe' where dni like '%$dni'<br>";
-			}
-			elseif (strlen ( $cargo_profe ) > "1" and !(is_numeric($cargo_profe))) {
-				$dni = substr ( $dni, 0, -2 );
-				$n_profe = mysqli_query($db_con, "select nombre from departamentos where dni='$dni'" );
-				$n_prof = mysqli_fetch_array ( $n_profe );
-				$unidad = $cargo_profe;
-				$n_tutor = mb_strtoupper ( $n_prof [0], 'UTF-8' );
+		foreach($_POST as $input_key => $input_value) {
+			
+			if ($input_key != "enviar" && ! empty($input_value)) {
 
-				mysqli_query($db_con, "insert INTO `FTUTORES` ( `unidad` , `tutor`, `observaciones1`, `observaciones2` ) VALUES ('$unidad', '$n_tutor', '', '')" );
+				//echo "$input_key => $input_value<br>";
 
-			} elseif (strlen ( $cargo_profe ) < "2") {
-				$dni=trim($dni);
-				mysqli_query($db_con, "update departamentos set cargo = ''" );
-				$dni = substr ($dni, 0, -1);
-				mysqli_query($db_con, "INSERT INTO `cargos` ( `dni` , `cargo` ) VALUES ('$dni', '$cargo_profe')" );
+				$input_value = trim($input_value);
+				$exp_dni = explode('_', $input_key);
+				$dni = trim($exp_dni[0]);
+
+				if ((! is_numeric($input_value)) && strlen($input_value) > 1) {
+					$resultNombreProfesor = mysqli_query($db_con, "SELECT nombre FROM departamentos WHERE dni = '$dni' LIMIT 1");
+					$rowNombreProfesor = mysqli_fetch_array($resultNombreProfesor);
+					$unidad = $input_value;
+					$n_tutor = $rowNombreProfesor[0];
+
+					mysqli_query($db_con, "INSERT INTO `FTUTORES` (`unidad` , `tutor`, `observaciones1`, `observaciones2`) VALUES ('$unidad', '$n_tutor', '', '')");
+
+				} 
+				elseif (strlen($input_value) < 2) {
+					$cargo = $input_value;
+					mysqli_query($db_con, "UPDATE departamentos SET cargo = '' WHERE dni = '$dni' LIMIT 1");
+					mysqli_query($db_con, "INSERT INTO `cargos` (`dni` , `cargo`) VALUES ('$dni', '$cargo')" );
+				}
 			}
 		}
 
-		$n_cargo = mysqli_query($db_con, "select dni from departamentos" );
-		while ( $n_carg = mysqli_fetch_array ( $n_cargo ) ) {
-			$num_cargos = "";
-			$num_car = mysqli_query($db_con, "select distinct cargo from cargos where dni = '$n_carg[0]'" );
-			while ( $num_carg = mysqli_fetch_array ( $num_car ) ) {
-				$num_cargos .= $num_carg [0];
+		$dniEmpleados = mysqli_query($db_con, "SELECT dni FROM departamentos");
+		while ($rowDniEmpleados = mysqli_fetch_array($dniEmpleados)) {
+			$serializeCargosEmpleado = "";
+			$cargosEmpleado = mysqli_query($db_con, "SELECT DISTINCT cargo FROM cargos WHERE dni = '".$rowDniEmpleados['dni']."'");
+			while ($rowCargosEmpleado = mysqli_fetch_array($cargosEmpleado)) {
+				$serializeCargosEmpleado .= $rowCargosEmpleado['cargo'];
 			}
-			mysqli_query($db_con, "update departamentos set cargo='$num_cargos' where dni='$n_carg[0]'" );
+			mysqli_query($db_con, "UPDATE departamentos SET cargo = '$serializeCargosEmpleado' WHERE dni = '".$rowDniEmpleados['dni']."' LIMIT 1");
 		}
-	echo '<div class="alert alert-success">
-	Los perfiles han sido asignados correctamente a los profesores.
-	          </div>';
+
+		echo '<div class="alert alert-success">Los perfiles han sido asignados correctamente a los profesores.</div>';
 	}
 	?>
 
@@ -102,15 +104,17 @@ include ("../menu.php");
 			<th><span data-bs="tooltip" title="Miembros del personal de Administracción y Servicios: Conserjes.">Conserje</span></th>
 			<th><span data-bs="tooltip" title="Miembros del personal de Administracción y Servicios: Administrativos">Administ.</span></th>
 			<th><span data-bs="tooltip" title="Todos los profesores que pertenecen al Equipo de Orientación, incluídos ATAL, Apoyo, PCPI, etc.">Orienta.</span></th>';
-		if($config['mod_bilingue']) $head .= '<th><span data-bs="tooltip" title="Profesores que participan en el Plan de Bilinguismo">Bilingüe</span></th>';
-
-		if ($config['mod_convivencia']==1) {
-		$head .= '<th><span data-bs="tooltip" title="Profesores encargados de atender a los alumnos en el Aula de Convivencia del Centro, si este cuenta con ella.">Conv.</span></th>';
+		if (isset($config['mod_bilingue']) && $config['mod_bilingue'] == 1) {
+			$head .= '<th><span data-bs="tooltip" title="Profesores que participan en el Plan de Bilinguismo">Bilingüe</span></th>';
 		}
-		if($config['mod_biblioteca']) $head .= '<th><span data-bs="tooltip" title="Profesores que participan en el Plan de Bibliotecas o se encargan de llevar la Biblioteca del Centro">Biblio.</span></th>';
+		if (isset($config['mod_convivencia']) && $config['mod_convivencia'] == 1) {
+			$head .= '<th><span data-bs="tooltip" title="Profesores encargados de atender a los alumnos en el Aula de Convivencia del Centro, si este cuenta con ella.">Conv.</span></th>';
+		}
+		if (isset($config['mod_biblioteca']) && $config['mod_biblioteca'] == 1) {
+			$head .= '<th><span data-bs="tooltip" title="Profesores que participan en el Plan de Bibliotecas o se encargan de llevar la Biblioteca del Centro">Biblio.</span></th>';
+		}
 		$head .= '<th><span data-bs="tooltip" title="Profesor encargado de las Relaciones de Género">Género</span></th>
 				  <th><span data-bs="tooltip" title="Departamento de Formacción, Innovación y Evaluación">DFEIE</span></th>
-				  <th><span data-bs="tooltip" title="Teléfono móvil del Profesor">Tfno. Móvil</span></th>
 				  <th>&nbsp;</th>
 			</tr>
 			</thead>';
@@ -126,14 +130,16 @@ include ("../menu.php");
 		<?php echo $head;?>
 			<tbody>
 		<?php
-		$carg0 = mysqli_query($db_con, "select distinct departamentos.nombre, departamentos.cargo, departamentos.dni, departamentos.idea, c_profes.telefono from departamentos, c_profes where departamentos.idea=c_profes.idea order by departamentos.nombre" );
-		$num_profes = mysqli_num_rows ( $carg0 );
-		while ( $carg1 = mysqli_fetch_array ( $carg0 ) ) {
-			$pro = $carg1 [0];
-			$car = $carg1 [1];
-			$dni = $carg1 [2];
-			$idea = $carg1 [3];
-			$telefono = $carg1 [4];
+		$resultDatosEmpleados = mysqli_query($db_con, "SELECT DISTINCT departamentos.nombre, departamentos.cargo, departamentos.dni, departamentos.idea, c_profes.telefono, c_profes.correo FROM departamentos JOIN c_profes ON departamentos.idea=c_profes.idea ORDER BY departamentos.nombre ASC");
+		$num_profes = mysqli_num_rows($resultDatosEmpleados);
+		$n_i = 0;
+		while($rowDatosEmpleados = mysqli_fetch_array($resultDatosEmpleados)) {
+			$pro = $rowDatosEmpleados['nombre'];
+			$car = $rowDatosEmpleados['cargo'];
+			$dni = $rowDatosEmpleados['dni'];
+			$idea = $rowDatosEmpleados['idea'];
+			$correo = $rowDatosEmpleados['correo'];
+			$telefono = $rowDatosEmpleados['telefono'];
 			if ($telefono=='0') {
 				$telefono="";
 			}
@@ -143,172 +149,95 @@ include ("../menu.php");
 			}
 			?>
 		<tr>
-				<td nowrap><small><?php
-			echo $pro;
-			?></small>
+			<td nowrap>
+				<strong><small><?php echo $pro;?></small></strong>
+				<?php if (! empty($telefono)): ?>
+				<p style="margin-bottom: 0; font-size: 0.75em;"><strong>Teléfono móvil:</strong> <?php echo $telefono; ?></p>
+				<?php endif; ?>
+				<?php if (! empty($correo)): ?>
+				<p style="margin-bottom: 0; font-size: 0.75em;"><?php echo $correo; ?></p>
+				<?php endif; ?>
 			</td>
 
 			<td class="text-center">
-				<input type="checkbox" name="<?php echo $dni; ?>z" value="z" <?php if (stristr ( $car, 'z' ) == TRUE) { echo "checked"; } if ($idea == "admin") { echo "checked disabled"; } ?> />
+				<input type="checkbox" name="<?php echo $dni; ?>_z" value="z" <?php echo (! empty($car) && strpos('z', $car) !== false) ? "checked" : ""; ?> <?php echo ($idea == "admin") ? "checked disabled" : ""; ?>  />
 			</td>
-
-				<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>1"
-					value="1" <?php
-			if (stristr ( $car, '1' ) == TRUE) {
-				echo "checked";
-			}
-			?>
-					 /></td>
-				<td class="form-inline" nowrap><input type="checkbox" name="<?php
-			echo $dni;
-			?>2"
-					value="2"
-					<?php
-			if (stristr ( $car, '2' ) == TRUE) {
-				echo "checked";
-			}
-			?> /> <select class="form-control input-sm" style="width: 100px;"
-					name="<?php
-			echo $dni;
-			?>2t">
-		  <?php
-			$curso_tut = mysqli_query($db_con, "select unidad from FTUTORES, departamentos where tutor=nombre and dni='$dni'" );
-			$curso_tut0 = mysqli_fetch_array ( $curso_tut );
-			$unidad = $curso_tut0 [0];
-			?>
-		  <option><?php
-		  if (strlen($unidad) > '1') {
-		  		echo $unidad;
-		  }
-			?></option>
-		<?php
-			echo "<option></option>";
-			$tipo = "select distinct unidad from alma order by unidad";
-			$tipo1 = mysqli_query($db_con, $tipo );
-			while ( $tipo2 = mysqli_fetch_array ( $tipo1 ) ) {
-				echo "<option value=\"" . $tipo2[0] . "\">" . $tipo2 [0] . "</option>";
-			}
-
-			?>
-		  </select></td>
-				<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>4"
-					value="4"
-					<?php
-			if (stristr ( $car, '4' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-				<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>15"
-					value="9"
-					<?php
-			if (stristr ( $car, '9' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-				<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>5"
-					value="5"
-					<?php
-			if (stristr ( $car, '5' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-				<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>6"
-					value="6"
-					<?php
-			if (stristr ( $car, '6' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-				<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>7"
-					value="7"
-					<?php
-			if (stristr ( $car, '7' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-				<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>8"
-					value="8"
-					<?php
-			if (stristr ( $car, '8' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-			<?php if($config['mod_bilingue']) { ?>
-				<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>9"
-					value="a"
-					<?php
-			if (stristr ( $car, 'a' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-			<?php } ?>
-			<?php if ($config['mod_convivencia']==1) { ?>
-			<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>10"
-					value="b"
-					<?php
-			if (stristr ( $car, 'b' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-			<?php } ?>
-
-			<?php if($config['mod_biblioteca']) { ?>
-			<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>11"
-					value="c"
-					<?php
-			if (stristr ( $car, 'c' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-			<?php } ?>
-			<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>12"
-					value="d"
-					<?php
-			if (stristr ( $car, 'd' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
-			<td class="text-center"><input type="checkbox" name="<?php
-			echo $dni;
-			?>13"
-					value="f"
-					<?php
-			if (stristr ( $car, 'f' ) == TRUE) {
-				echo "checked";
-			}
-			?> /></td>
 
 			<td class="text-center">
-			<input type="tel" class="form-control input-sm" name="<?php echo $dni; ?>14" value="<?php echo $telefono; ?>" style="width:100px" min="9" max="9"/>
+				<?php if ($idea == "admin"): ?>
+				<input type="checkbox" name="" value="1" <?php echo ($idea == "admin") ? "checked disabled" : ""; ?> />
+				<input type="hidden" name="<?php echo $dni; ?>_1" value="1">
+				<?php else: ?>
+				<input type="checkbox" name="<?php echo $dni; ?>_1" value="1" <?php echo (! empty($car) && strpos('1', $car) !== false) ? "checked" : ""; ?> <?php echo ($idea == "admin") ? "checked disabled" : ""; ?> />
+				<?php endif; ?>
+			</td>
+			<td class="form-inline" nowrap>
+				<input type="checkbox" name="<?php echo $dni; ?>_2" value="2" <?php echo (! empty($car) && strpos('2', $car) !== false) ? "checked" : ""; ?> /> 
+				<select class="form-control input-sm" style="width: auto;" name="<?php echo $dni; ?>_2t">
+					<option value=""></option>
+					<?php
+					$tutorUnidadCentro = mysqli_query($db_con, "SELECT unidad FROM FTUTORES WHERE tutor = '".$pro."' LIMIT 1");
+					$rowTutorUnidadCentro = mysqli_fetch_array($tutorUnidadCentro);
+					$tutorUnidad = $rowTutorUnidadCentro['unidad'];
+
+					$unidadesCentro = mysqli_query($db_con, "SELECT nomunidad FROM unidades ORDER BY idcurso ASC, nomunidad ASC");
+					while($rowUnidadesCentro = mysqli_fetch_array($unidadesCentro)) {
+					?>
+			  		<option value="<?php echo $rowUnidadesCentro['nomunidad']; ?>" <?php echo ($tutorUnidad == $rowUnidadesCentro['nomunidad']) ? "selected" : ""; ?>><?php echo $rowUnidadesCentro['nomunidad']; ?></option>
+					<?php } ?>
+		  		</select>
+		  	</td>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_4" value="4" <?php echo (! empty($car) && strpos('4', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_15" value="9" <?php echo (! empty($car) && strpos('9', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_5" value="5" <?php echo (! empty($car) && strpos('5', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_6" value="6" <?php echo (! empty($car) && strpos('6', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_7" value="7" <?php echo (! empty($car) && strpos('7', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_8" value="8" <?php echo (! empty($car) && strpos('8', $car) !== false) ? "checked" : ""; ?> />
 			</td>
 
-			<td class="text-center"><a href="cargos.php?borrar=1&dni_profe=<?php echo $dni;?>" data-bb='confirm-delete'><span class="far fa-trash-alt fa-lg fa-fw"></span></a></td>
-			</tr>
-		<?php
-			}
-		?>
+			<?php if (isset($config['mod_bilingue']) && $config['mod_bilingue'] == 1): ?>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_9" value="a" <?php echo (! empty($car) && strpos('a', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+			<?php endif; ?>
+
+			<?php if (isset($config['mod_convivencia']) && $config['mod_convivencia'] == 1): ?>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_10" value="b" <?php echo (! empty($car) && strpos('b', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+			<?php endif; ?>
+
+			<?php if (isset($config['mod_biblioteca']) && $config['mod_biblioteca'] == 1): ?>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_11" value="c" <?php echo (! empty($car) && strpos('c', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+			<?php endif; ?>
+
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_12" value="d" <?php echo (! empty($car) && strpos('d', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+			<td class="text-center">
+				<input type="checkbox" name="<?php echo $dni; ?>_13" value="f" <?php echo (! empty($car) && strpos('f', $car) !== false) ? "checked" : ""; ?> />
+			</td>
+
+			<td class="text-center">
+				<?php if ($idea != "admin"): ?>
+				<a href="cargos.php?borrar=1&dni_profe=<?php echo $dni; ?>" data-bb='confirm-delete'><span class="far fa-trash-alt fa-lg fa-fw"></span></a>
+				<?php endif; ?>
+			</td>
+		</tr>
+		<?php } ?>
 		</tbody>
 		</table>
